Move integration tests to root folder
