+++ test/refs_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/grafana/nanogit/client"
-	"github.com/grafana/nanogit/client/integration/helpers"
+	"github.com/grafana/nanogit/test/helpers"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
